moved RichTextPreview to a separate file
