+++ packages/bodiless-richtext/src/RichText.tsx
@@ -28,9 +28,6 @@ import {
 } from '@bodiless/core';
 import {
   designable,
-  replaceWith,
-  flowIf,
-  hasProp,
   H1,
   H2,
   H3,
@@ -44,7 +41,6 @@ import {
   extendDesign,
   Design,
   DesignableComponents,
-  addProps,
 } from '@bodiless/fclasses';
 import {
   withSlateEditor,
@@ -60,10 +56,6 @@ import {
   getSchema,
   getSelectorButtons,
 } from './RichTextItemGetters';
-import { withId, asMark } from './RichTextItemSetters';
-import {
-  RichTextComponents, RichTextComponent,
-} from './Type';
 import TextSelectorButton from './components/TextSelectorButton';
 import { uiContext, getUI, UI } from './RichTextContext';
 import defaultValue from './default-value';
@@ -81,7 +73,9 @@ import {
   withHeader2Meta,
   withHeader3Meta,
 } from './meta';
-import PluginButton from './components/PluginButton';
+import withDefaults from './withDefaults';
+import { withPreview } from './RichTextPreview';
+import type { RichTextProps } from './Type';
 
 type WithSlateSchemaTypeProps = {
   schema: object,
@@ -176,13 +170,6 @@ const RichTextProvider = flowRight(
   withSlateSchema,
 )(React.Fragment) as RichTextProviderType;
 
-export type RichTextProps<P> = {
-  components: DesignableComponents,
-  ui?: UI,
-  initialValue?: object,
-  nodeKey?: string,
-};
-
 /**
  * @private
  * Observer wrapper around hover menu which hides it when not in edit mode.
@@ -195,22 +182,6 @@ const EditOnlyHoverMenu$: FC<Pick<Required<UI>, 'HoverMenu'>> = ({ HoverMenu, ch
 };
 const EditOnlyHoverMenu = observer(EditOnlyHoverMenu$);
 
-/**
- * ensure the componets have a type (we default to mark) as well as ensuring there is an id
- * @param components which set of component on which we should operate
- */
-const withDefaults = (components: DesignableComponents) => {
-  const withDefaultType = (Component: ComponentType<any>) => (
-    // eslint-disable-next-line no-prototype-builtins
-    Component.hasOwnProperty('type') ? Component : asMark(Component)
-  );
-  return Object.getOwnPropertyNames(components).reduce(
-    (acc, id) => (
-      { ...acc, [id]: flow(withDefaultType, withId(id))(acc[id]) as RichTextComponent }
-    ),
-    components,
-  ) as RichTextComponents;
-};
 const BasicRichText = <P extends object, D extends object>(props: P & RichTextProps<D>) => {
   const {
     initialValue,
@@ -305,36 +276,9 @@ const apply = (design: Design<DesignableComponents>) => {
   return applyDesign(start)(extendDesign(finalDesign)(design));
 };
 
-const BaseRichTextPreview = <P extends object, D extends object>(props: P & RichTextProps<D>) => {
-  const { components, ui } = props;
-  const finalComponents = withDefaults(components);
-  const { PreviewWrapper } = getUI(ui);
-  return (
-    <PreviewWrapper>
-      {
-        Object.values(finalComponents)
-          // eslint-disable-next-line react/no-array-index-key
-          .map((C, i) => C.hoverButton && <PluginButton key={i} icon={C.hoverButton.icon} componentName="Button" />)
-      }
-    </PreviewWrapper>
-  );
-};
-
-const asPreview = addProps({
-  preview: 1,
-});
-
-const withPreview = flowIf(hasProp('preview'))(
-  withoutProps(['preview']),
-  replaceWith(BaseRichTextPreview),
-);
-
 const RichText = flow(
   withPreview,
   designable(apply),
 )(BasicRichText);
 
 export default RichText;
-export {
-  asPreview,
-};
