+++ packages/bodiless-richtext/src/index.ts
@@ -14,9 +14,8 @@
 
 export {
   default as RichText,
-  asPreview,
 } from './RichText';
-export type { RichTextProps } from './RichText';
+export { asPreview } from './RichTextPreview';
 export { default as asFloat } from './Float';
 export * from './RichTextItemGetters';
 export * from './RichTextItemSetters';
